Accept IPv6 CIDR prefixes, too
When accepting CIDR prefixes from the user - in security groups and when
adding a subnet to a network - the code uses regular expressions to
check for validity of these prefixes.

Before this patch, the regular expressions would only match IPv4
prefixes.  This patch modifies them to also accept IPv6 prefixes.  In
addition, the subnet creation code will automatically derive the correct
value of the ip_version from the kind of prefix that the user has
entered.